+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 * Unreleased
+    * Add `tstrings.h` which adds various `strxxx_T()` overloaded functions
+      which take either a `const char*` or a `const __FlashStringHelper*`
+      parameter.
+        * Allows writing C++ template code which is agnostic to whether the
+          string argument is in normal memory or flash memory. The compiler will
+          select the specific implementation automatically at compile-time.
 * 1.6.1 (2023-06-09)
     * Fix backwards compatibility breakage in `KString()` constructor.
         * v1.5.2 supported only: